fix: 0xFF byte literal in GeoAssetIndexTest needed explicit .toByte()
Kotlin does not narrow 0xFF (255) into a Byte for byteArrayOf(...), so the
corruption fixture failed to compile the unit-test source set with
'Argument type mismatch: actual type is 'Int', but 'Byte' was expected'
(GeoAssetIndexTest.kt:174). All other byte literals in the file fit in a
Byte and were audited; the production source set compiled clean.

+++ app/src/test/java/com/marbleng/app/core/GeoAssetIndexTest.kt
@@ -171,7 +171,7 @@ class GeoAssetIndexTest {
         val site = geositeFile(geoSite("google", listOf(domain(2, "google.com"))))
         val dir = tmp.newFolder()
         File(dir, "geosite.dat").writeBytes(site + byteArrayOf(0x7F, 0x2A, 0x11)) // junk tail
-        File(dir, "geoip.dat").writeBytes(byteArrayOf(0xFF, 0xFF.toByte(), 0x01))
+        File(dir, "geoip.dat").writeBytes(byteArrayOf(0xFF.toByte(), 0xFF.toByte(), 0x01))
 
         val snapshot = GeoAssetIndex.update(dir)
         // The well-formed entries still landed; the junk ended the scan where it started.
